changed dashboard popups to parallel rouiting
