--- components/HomePage/Arbswap/Dashboard/Form.tsx
@@ -1,9 +1,0 @@
-'use client'
-
-import React from 'react'
-
-function Form() {
-  return <div>form</div>
-}
-
-export default Form
